Fixed bug with periodic longitudes
It now get the boundaries before assigning, otherwise it grabs the added eastern boundary in the first step to the western boundary added in the second step, which leads to repeated boundaries. It also add only a single column (lon) instead of the entire domain, which produces repeated longitudes and longitudes that are not monotically increasing.